+++ src/data/symbol_svgs.json
