+++ src/util/svg.ts
@@ -1,6 +1,6 @@
-import symbolPaths from "../symbol_svgs.json";
-import countryPaths from "../country_outlines.json";
-import socialPaths from "../social_svgs.json";
+import symbolPaths from "../data/symbol_svgs.json";
+import countryPaths from "../data/country_outlines.json";
+import socialPaths from "../data/social_svgs.json";
 
 export function getPath(name: string) {
   const allPaths = [...countryPaths, ...symbolPaths, ...socialPaths];
